A Robocode robot.
As an end-of-the-year project, everyone in my class made a robot in Robocode, which we then battled each other in a March Madness-style tournament to crown a champion. Covfefe was not that champion, but nevertheless he earned a respectable third place. Covfefe was named after one of the (many) legendary tweets from the 45th President of the United States. This was originally programmed in the spring of 2017, but I'd like to go back sometime and make some improvements.